Fix verifyProcedure
There was a need to find out if a solver can support non-SMT-LIB
commands, so I subtyped Cmd with ExtCmd and Solver with ExtSolver.

First, I tried to make Cmd a subtype of ExtCmd and parameterize Solver
with its kind of command. This would be correct IMO since then Cmd could
be assigned to ExtCmd but not vice versa. However, I couldn't make that
work without a _lot_ of rewriting.